5.10.147 first commit

+++ drivers/mtd/maps/physmap-bt1-rom.c
@@ -14,6 +14,7 @@
 #include <linux/mtd/xip.h>
 #include <linux/mux/consumer.h>
 #include <linux/of.h>
+#include <linux/of_device.h>
 #include <linux/platform_device.h>
 #include <linux/string.h>
 #include <linux/types.h>
@@ -30,12 +31,12 @@ static map_word __xipram bt1_rom_map_read(struct map_info *map,
 					  unsigned long ofs)
 {
 	void __iomem *src = map->virt + ofs;
-	unsigned int shift;
+	unsigned long shift;
 	map_word ret;
 	u32 data;
 
 	/* Read data within offset dword. */
-	shift = (uintptr_t)src & 0x3;
+	shift = (unsigned long)src & 0x3;
 	data = readl_relaxed(src - shift);
 	if (!shift) {
 		ret.x[0] = data;
@@ -59,7 +60,7 @@ static void __xipram bt1_rom_map_copy_from(struct map_info *map,
 					   ssize_t len)
 {
 	void __iomem *src = map->virt + from;
-	unsigned int shift, chunk;
+	ssize_t shift, chunk;
 	u32 data;
 
 	if (len <= 0 || from >= map->size)
@@ -74,7 +75,7 @@ static void __xipram bt1_rom_map_copy_from(struct map_info *map,
 	 * up into the next three stages: unaligned head, aligned body,
 	 * unaligned tail.
 	 */
-	shift = (uintptr_t)src & 0x3;
+	shift = (ssize_t)src & 0x3;
 	if (shift) {
 		chunk = min_t(ssize_t, 4 - shift, len);
 		data = readl_relaxed(src - shift);
